LayoutTests/imported/w3c/web-platform-tests/css/css-text/text-transform/math/text-transform-math-auto-003-expected.txt


FAIL Selection.toString() for math-auto 'A' returns the transformed character. assert_equals: expected "ğ´" but got "A"
FAIL Selection.toString() for math-auto 'B' returns the transformed character. assert_equals: expected "ğµ" but got "B"
FAIL Selection.toString() for math-auto 'C' returns the transformed character. assert_equals: expected "ğ¶" but got "C"
FAIL Selection.toString() for math-auto 'D' returns the transformed character. assert_equals: expected "ğ·" but got "D"
FAIL Selection.toString() for math-auto 'E' returns the transformed character. assert_equals: expected "ğ¸" but got "E"
FAIL Selection.toString() for math-auto 'F' returns the transformed character. assert_equals: expected "ğ¹" but got "F"
FAIL Selection.toString() for math-auto 'G' returns the transformed character. assert_equals: expected "ğº" but got "G"
FAIL Selection.toString() for math-auto 'H' returns the transformed character. assert_equals: expected "ğ»" but got "H"
FAIL Selection.toString() for math-auto 'I' returns the transformed character. assert_equals: expected "ğ¼" but got "I"
FAIL Selection.toString() for math-auto 'J' returns the transformed character. assert_equals: expected "ğ½" but got "J"
FAIL Selection.toString() for math-auto 'K' returns the transformed character. assert_equals: expected "ğ¾" but got "K"
FAIL Selection.toString() for math-auto 'L' returns the transformed character. assert_equals: expected "ğ¿" but got "L"
FAIL Selection.toString() for math-auto 'M' returns the transformed character. assert_equals: expected "ğ‘€" but got "M"
FAIL Selection.toString() for math-auto 'N' returns the transformed character. assert_equals: expected "ğ‘" but got "N"
FAIL Selection.toString() for math-auto 'O' returns the transformed character. assert_equals: expected "ğ‘‚" but got "O"
FAIL Selection.toString() for math-auto 'P' returns the transformed character. assert_equals: expected "ğ‘ƒ" but got "P"
FAIL Selection.toString() for math-auto 'Q' returns the transformed character. assert_equals: expected "ğ‘„" but got "Q"
FAIL Selection.toString() for math-auto 'R' returns the transformed character. assert_equals: expected "ğ‘…" but got "R"
FAIL Selection.toString() for math-auto 'S' returns the transformed character. assert_equals: expected "ğ‘†" but got "S"
FAIL Selection.toString() for math-auto 'T' returns the transformed character. assert_equals: expected "ğ‘‡" but got "T"
FAIL Selection.toString() for math-auto 'U' returns the transformed character. assert_equals: expected "ğ‘ˆ" but got "U"
FAIL Selection.toString() for math-auto 'V' returns the transformed character. assert_equals: expected "ğ‘‰" but got "V"
FAIL Selection.toString() for math-auto 'W' returns the transformed character. assert_equals: expected "ğ‘Š" but got "W"
FAIL Selection.toString() for math-auto 'X' returns the transformed character. assert_equals: expected "ğ‘‹" but got "X"
FAIL Selection.toString() for math-auto 'Y' returns the transformed character. assert_equals: expected "ğ‘Œ" but got "Y"
FAIL Selection.toString() for math-auto 'Z' returns the transformed character. assert_equals: expected "ğ‘" but got "Z"
FAIL Selection.toString() for math-auto 'a' returns the transformed character. assert_equals: expected "ğ‘" but got "a"
FAIL Selection.toString() for math-auto 'b' returns the transformed character. assert_equals: expected "ğ‘" but got "b"
FAIL Selection.toString() for math-auto 'c' returns the transformed character. assert_equals: expected "ğ‘" but got "c"
FAIL Selection.toString() for math-auto 'd' returns the transformed character. assert_equals: expected "ğ‘‘" but got "d"
FAIL Selection.toString() for math-auto 'e' returns the transformed character. assert_equals: expected "ğ‘’" but got "e"
FAIL Selection.toString() for math-auto 'f' returns the transformed character. assert_equals: expected "ğ‘“" but got "f"
FAIL Selection.toString() for math-auto 'g' returns the transformed character. assert_equals: expected "ğ‘”" but got "g"
FAIL Selection.toString() for math-auto 'h' returns the transformed character. assert_equals: expected "â„" but got "h"
FAIL Selection.toString() for math-auto 'i' returns the transformed character. assert_equals: expected "ğ‘–" but got "i"
FAIL Selection.toString() for math-auto 'j' returns the transformed character. assert_equals: expected "ğ‘—" but got "j"
FAIL Selection.toString() for math-auto 'k' returns the transformed character. assert_equals: expected "ğ‘˜" but got "k"
FAIL Selection.toString() for math-auto 'l' returns the transformed character. assert_equals: expected "ğ‘™" but got "l"
FAIL Selection.toString() for math-auto 'm' returns the transformed character. assert_equals: expected "ğ‘š" but got "m"
FAIL Selection.toString() for math-auto 'n' returns the transformed character. assert_equals: expected "ğ‘›" but got "n"
FAIL Selection.toString() for math-auto 'o' returns the transformed character. assert_equals: expected "ğ‘œ" but got "o"
FAIL Selection.toString() for math-auto 'p' returns the transformed character. assert_equals: expected "ğ‘" but got "p"
FAIL Selection.toString() for math-auto 'q' returns the transformed character. assert_equals: expected "ğ‘" but got "q"
FAIL Selection.toString() for math-auto 'r' returns the transformed character. assert_equals: expected "ğ‘Ÿ" but got "r"
FAIL Selection.toString() for math-auto 's' returns the transformed character. assert_equals: expected "ğ‘ " but got "s"
FAIL Selection.toString() for math-auto 't' returns the transformed character. assert_equals: expected "ğ‘¡" but got "t"
FAIL Selection.toString() for math-auto 'u' returns the transformed character. assert_equals: expected "ğ‘¢" but got "u"
FAIL Selection.toString() for math-auto 'v' returns the transformed character. assert_equals: expected "ğ‘£" but got "v"
FAIL Selection.toString() for math-auto 'w' returns the transformed character. assert_equals: expected "ğ‘¤" but got "w"
FAIL Selection.toString() for math-auto 'x' returns the transformed character. assert_equals: expected "ğ‘¥" but got "x"
FAIL Selection.toString() for math-auto 'y' returns the transformed character. assert_equals: expected "ğ‘¦" but got "y"
FAIL Selection.toString() for math-auto 'z' returns the transformed character. assert_equals: expected "ğ‘§" but got "z"
FAIL Selection.toString() for math-auto 'Ä±' returns the transformed character. assert_equals: expected "ğš¤" but got "Ä±"
FAIL Selection.toString() for math-auto 'È·' returns the transformed character. assert_equals: expected "ğš¥" but got "È·"
FAIL Selection.toString() for math-auto 'Î‘' returns the transformed character. assert_equals: expected "ğ›¢" but got "Î‘"
FAIL Selection.toString() for math-auto 'Î’' returns the transformed character. assert_equals: expected "ğ›£" but got "Î’"
FAIL Selection.toString() for math-auto 'Î“' returns the transformed character. assert_equals: expected "ğ›¤" but got "Î“"
FAIL Selection.toString() for math-auto 'Î”' returns the transformed character. assert_equals: expected "ğ›¥" but got "Î”"
FAIL Selection.toString() for math-auto 'Î•' returns the transformed character. assert_equals: expected "ğ›¦" but got "Î•"
FAIL Selection.toString() for math-auto 'Î–' returns the transformed character. assert_equals: expected "ğ›§" but got "Î–"
FAIL Selection.toString() for math-auto 'Î—' returns the transformed character. assert_equals: expected "ğ›¨" but got "Î—"
FAIL Selection.toString() for math-auto 'Î˜' returns the transformed character. assert_equals: expected "ğ›©" but got "Î˜"
FAIL Selection.toString() for math-auto 'Î™' returns the transformed character. assert_equals: expected "ğ›ª" but got "Î™"
FAIL Selection.toString() for math-auto 'Îš' returns the transformed character. assert_equals: expected "ğ›«" but got "Îš"
FAIL Selection.toString() for math-auto 'Î›' returns the transformed character. assert_equals: expected "ğ›¬" but got "Î›"
FAIL Selection.toString() for math-auto 'Îœ' returns the transformed character. assert_equals: expected "ğ›­" but got "Îœ"
FAIL Selection.toString() for math-auto 'Î' returns the transformed character. assert_equals: expected "ğ›®" but got "Î"
FAIL Selection.toString() for math-auto 'Î' returns the transformed character. assert_equals: expected "ğ›¯" but got "Î"
FAIL Selection.toString() for math-auto 'ÎŸ' returns the transformed character. assert_equals: expected "ğ›°" but got "ÎŸ"
FAIL Selection.toString() for math-auto 'Î ' returns the transformed character. assert_equals: expected "ğ›±" but got "Î "
FAIL Selection.toString() for math-auto 'Î¡' returns the transformed character. assert_equals: expected "ğ›²" but got "Î¡"
FAIL Selection.toString() for math-auto 'Î£' returns the transformed character. assert_equals: expected "ğ›´" but got "Î£"
FAIL Selection.toString() for math-auto 'Î¤' returns the transformed character. assert_equals: expected "ğ›µ" but got "Î¤"
FAIL Selection.toString() for math-auto 'Î¥' returns the transformed character. assert_equals: expected "ğ›¶" but got "Î¥"
FAIL Selection.toString() for math-auto 'Î¦' returns the transformed character. assert_equals: expected "ğ›·" but got "Î¦"
FAIL Selection.toString() for math-auto 'Î§' returns the transformed character. assert_equals: expected "ğ›¸" but got "Î§"
FAIL Selection.toString() for math-auto 'Î¨' returns the transformed character. assert_equals: expected "ğ›¹" but got "Î¨"
FAIL Selection.toString() for math-auto 'Î©' returns the transformed character. assert_equals: expected "ğ›º" but got "Î©"
FAIL Selection.toString() for math-auto 'Î±' returns the transformed character. assert_equals: expected "ğ›¼" but got "Î±"
FAIL Selection.toString() for math-auto 'Î²' returns the transformed character. assert_equals: expected "ğ›½" but got "Î²"
FAIL Selection.toString() for math-auto 'Î³' returns the transformed character. assert_equals: expected "ğ›¾" but got "Î³"
FAIL Selection.toString() for math-auto 'Î´' returns the transformed character. assert_equals: expected "ğ›¿" but got "Î´"
FAIL Selection.toString() for math-auto 'Îµ' returns the transformed character. assert_equals: expected "ğœ€" but got "Îµ"
FAIL Selection.toString() for math-auto 'Î¶' returns the transformed character. assert_equals: expected "ğœ" but got "Î¶"
FAIL Selection.toString() for math-auto 'Î·' returns the transformed character. assert_equals: expected "ğœ‚" but got "Î·"
FAIL Selection.toString() for math-auto 'Î¸' returns the transformed character. assert_equals: expected "ğœƒ" but got "Î¸"
FAIL Selection.toString() for math-auto 'Î¹' returns the transformed character. assert_equals: expected "ğœ„" but got "Î¹"
FAIL Selection.toString() for math-auto 'Îº' returns the transformed character. assert_equals: expected "ğœ…" but got "Îº"
FAIL Selection.toString() for math-auto 'Î»' returns the transformed character. assert_equals: expected "ğœ†" but got "Î»"
FAIL Selection.toString() for math-auto 'Î¼' returns the transformed character. assert_equals: expected "ğœ‡" but got "Î¼"
FAIL Selection.toString() for math-auto 'Î½' returns the transformed character. assert_equals: expected "ğœˆ" but got "Î½"
FAIL Selection.toString() for math-auto 'Î¾' returns the transformed character. assert_equals: expected "ğœ‰" but got "Î¾"
FAIL Selection.toString() for math-auto 'Î¿' returns the transformed character. assert_equals: expected "ğœŠ" but got "Î¿"
FAIL Selection.toString() for math-auto 'Ï€' returns the transformed character. assert_equals: expected "ğœ‹" but got "Ï€"
FAIL Selection.toString() for math-auto 'Ï' returns the transformed character. assert_equals: expected "ğœŒ" but got "Ï"
FAIL Selection.toString() for math-auto 'Ï‚' returns the transformed character. assert_equals: expected "ğœ" but got "Ï‚"
FAIL Selection.toString() for math-auto 'Ïƒ' returns the transformed character. assert_equals: expected "ğœ" but got "Ïƒ"
FAIL Selection.toString() for math-auto 'Ï„' returns the transformed character. assert_equals: expected "ğœ" but got "Ï„"
FAIL Selection.toString() for math-auto 'Ï…' returns the transformed character. assert_equals: expected "ğœ" but got "Ï…"
FAIL Selection.toString() for math-auto 'Ï†' returns the transformed character. assert_equals: expected "ğœ‘" but got "Ï†"
FAIL Selection.toString() for math-auto 'Ï‡' returns the transformed character. assert_equals: expected "ğœ’" but got "Ï‡"
FAIL Selection.toString() for math-auto 'Ïˆ' returns the transformed character. assert_equals: expected "ğœ“" but got "Ïˆ"
FAIL Selection.toString() for math-auto 'Ï‰' returns the transformed character. assert_equals: expected "ğœ”" but got "Ï‰"
FAIL Selection.toString() for math-auto 'Ï‘' returns the transformed character. assert_equals: expected "ğœ—" but got "Ï‘"
FAIL Selection.toString() for math-auto 'Ï•' returns the transformed character. assert_equals: expected "ğœ™" but got "Ï•"
FAIL Selection.toString() for math-auto 'Ï–' returns the transformed character. assert_equals: expected "ğœ›" but got "Ï–"
FAIL Selection.toString() for math-auto 'Ï°' returns the transformed character. assert_equals: expected "ğœ˜" but got "Ï°"
FAIL Selection.toString() for math-auto 'Ï±' returns the transformed character. assert_equals: expected "ğœš" but got "Ï±"
FAIL Selection.toString() for math-auto 'Ï´' returns the transformed character. assert_equals: expected "ğ›³" but got "Ï´"
FAIL Selection.toString() for math-auto 'Ïµ' returns the transformed character. assert_equals: expected "ğœ–" but got "Ïµ"
FAIL Selection.toString() for math-auto 'âˆ‚' returns the transformed character. assert_equals: expected "ğœ•" but got "âˆ‚"
FAIL Selection.toString() for math-auto 'âˆ‡' returns the transformed character. assert_equals: expected "ğ›»" but got "âˆ‡"
